requirements.txt
# ProjectForge Dependencies
# AI-Powered Business Analysis Tool

# Core Framework
crewai>=0.70.0

# LLM Integration
google-generativeai>=0.8.0

# Environment Configuration
python-dotenv>=1.0.0

# Markdown Processing
markdown>=3.7

# PDF Generation
weasyprint>=63.0

# Web UI
streamlit>=1.31.0

# Note: WeasyPrint requires system-level dependencies on macOS:
# brew install pango gdk-pixbuf libffi
